Helpers: add cient wrapper.
This adds a wrapper client that knows how to get a set of credentials
specifically for what we've asked.

The current API implementation always returns a full list of
credentials, meaning that we need to filter out credentials if needed if
people specify a `credentials` block in their configuration.

This also knows how to handle missing values and match their default
values.